requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-groq
pydantic
python-multipart
jinja2
aiofiles
pytest
httpx
python-jose
passlib
bcrypt